src/lambda-Textract/requirements.txt
boto3
amazon-textract-response-parser
amazon-textract-caller
amazon-textract-prettyprinter
amazon-textract-textractor